refactor: Move event emitter to services
